Handle cases where getPlayerSlot() returns undefined
When bancho.js joins an existing lobby, you can receive playerMoved or
playerLeft events *before* bancho.js had the time to fetch the lobby
info. When that happens, an unhandled error is thrown, since "undefined"
is not a valid key for the slots array.

I chose to just check if getPlayerSlot() returns something, but a better
fix might be to ignore those events until the lobby is initialized.